Consolidate Penumbra directory configuration
